Move all shared selectors into main selectors directory
