+++ fronts-client/src/reducers/groupsReducer.ts
@@ -1,7 +1,7 @@
 import { Action } from '../shared/types/Action';
-import { insertAndDedupeSiblings } from '../util/insertAndDedupeSiblings';
+import { insertAndDedupeSiblings } from 'util/insertAndDedupeSiblings';
 import { State } from 'reducers/sharedReducer';
-import { selectCards, selectGroupSiblings } from 'shared/selectors/shared';
+import { selectCards, selectGroupSiblings } from 'selectors/shared';
 import { capGroupCards } from 'util/capGroupCards';
 import keyBy from 'lodash/keyBy';
