+++ fronts-client/src/selectors/__tests__/shared.spec.ts
