+++ fronts-client/src/selectors/collection.ts
