+++ fronts-client/src/shared/components/article/Article.tsx
@@ -7,7 +7,7 @@ import {
   createSelectArticleFromCard,
   selectSharedState,
   selectCard
-} from '../../selectors/shared';
+} from '../../../selectors/shared';
 import { selectors } from 'bundles/externalArticlesBundle';
 import { State } from '../../types/State';
 import { DerivedArticle } from '../../../types/Article';
