+++ fronts-client/src/util/shared.ts
@@ -5,7 +5,7 @@ import {
   Card
 } from 'shared/types/Collection';
 import { selectors as collectionSelectors } from 'bundles/collectionsBundle';
-import { selectSharedState } from 'shared/selectors/shared';
+import { selectSharedState } from 'selectors/shared';
 import { State } from 'types/State';
 
 import { normalize, denormalize } from './schema';
